Move content query processing code in to separate file
